Options added, StemmerApplier expanded, refactoring
- DeimosConfig updated to add certain booleans that control certain
functions, for example:
1. An option to check if output text files should be deleted after
zipping
2. An option to choose between simple and hashed file names (new:
Optionally, have less crazy filenames.
3. An option to choose between sample XML dmoz and full data.

- Porter-Stemming (StemmerApplier) was updated: it now contains two
simple functions to return the stemmed output for a single word and a
text containing multiple words. (some more changes to optimize were
done)

- Added a few more comments, renamed some other classes and organized
other code.

+++ src/deimos/phase1/ExportAll.java
@@ -10,7 +10,7 @@ import org.sqlite.SQLiteException;
 
 import deimos.common.DeimosConfig;
 import deimos.common.Mailer;
-import deimos.common.ProcessUtils;
+import deimos.common.ProcessFileUtils;
 
 /**
  * Combines all export functions,
@@ -26,6 +26,7 @@ import deimos.common.ProcessUtils;
 public class ExportAll {
 	
 	public static final String PROCESS_CHROME_WIN = "chrome.exe";
+
 	
 	public static boolean killChrome() {
 		System.out.print("Ensuring Chrome is not running: ");
@@ -34,9 +35,9 @@ public class ExportAll {
 		
 		try
 		{
-			if(ProcessUtils.isProcessRunning(PROCESS_CHROME_WIN))
+			if(ProcessFileUtils.isProcessRunning(PROCESS_CHROME_WIN))
 			{
-				ProcessUtils.killProcess(PROCESS_CHROME_WIN);
+				ProcessFileUtils.killProcess(PROCESS_CHROME_WIN);
 				killed = true;
 			}
 			else {
@@ -123,7 +124,8 @@ public class ExportAll {
 
        	System.out.println("Finished!");
 		
-		deleteOutputFiles();
+       	if(DeimosConfig.OPTION_DELETE_P1_OUTPUT)
+       		deleteOutputFiles();
 	}
 
 }
